feat: ship external import setup and quality profile upgrades
Add setup and settings flows for external import and download clients.
Expand quality profile scoring, release-group rules, and release parsing.
Improve media info extraction and related backend/frontend integration paths.

+++ crates/scryer-application/src/app_usecase_admin.rs
@@ -92,7 +92,7 @@ impl AppUseCase {
 
             match nix::sys::statvfs::statvfs(path.as_str()) {
                 Ok(stat) => {
-                    let block_size = stat.block_size() as u64;
+                    let block_size = stat.block_size();
                     let total = stat.blocks() as u64 * block_size;
                     let free = stat.blocks_available() as u64 * block_size;
                     let used = total.saturating_sub(free);
